Elm-svelte-app: Add initial value to elmBridge
When svelte runs store subscriptions the first time it expects the
subscription function to be called with a value. Elm does not
run subscriptions when they are first subscribed, so svelte will
interperet it as undefined the first time it gets called. Calling
the subscription function manually the first time solves this.